feat(controller): support vmnetcfg disabled mode
When an VirtualMachineNetworkConfig is disabled, it will free up the
slot it occupied in IPAM and remove the MAC cache entry. Also, the
referencing IPPool's status will be updated. However, the allocation
record still remain on its status field. When the
VirtualMachineNetworkConfig is enabled again, the same IP address will
be allocated.

Signed-off-by: Zespre Chang <zespre.chang@suse.com>