feat(ippool): add serverIP, start, and end fields if missing
If IPPool object is created without explicitly specified the serverIP,
the start and end IP of the pool, the IPPool mutator will try to assign
IPs to them. Also, if the serverIP or router is within the pool range,
i.e., in between the start and end IPs, those IPs will then be marked
as "RESERVED" (by the controller) and will not be allocatable.

Signed-off-by: Zespre Chang <zespre.chang@suse.com>